Workout Mern stack app
Continue the implementation of signup feature
we want the user to be able to signup with its email and password.
The password needs to be hashed
If the email provided is already in use, the application must throw an error saying 'User already in use'
We also want to validate the email and password before processing them and create the user with them.
If the email and the password are not provided the application must throw an error saying 'Please provide an email and password'
If the email is not a valid email we throw an error saying 'Email is not valid' as well
If the password is not strong we throw an error saying 'Password not strong enough'
We implemented these steps in a static method that we added the userSchema.